+++ docroot/modules/custom/mass_org_access/mass_org_access.services.yml
@@ -34,7 +34,6 @@ services:
       - '@mass_org_access.org_access_checker'
       - '@messenger'
       - '@mass_org_access.settings'
-      - '@current_user'
     tags:
       - { name: drupal.hook }
