+++ docroot/modules/custom/mass_org_access/src/Hook/MassOrgAccessHooks.php
@@ -13,7 +13,6 @@ use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Hook\Attribute\Hook;
 use Drupal\Core\Messenger\MessengerInterface;
 use Drupal\Core\Session\AccountInterface;
-use Drupal\Core\Session\AccountProxyInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\mass_org_access\OrgAccessChecker;
 use Drupal\mass_org_access\OrgAccessSettings;
@@ -31,7 +30,6 @@ class MassOrgAccessHooks {
     private readonly OrgAccessChecker $orgAccessChecker,
     private readonly MessengerInterface $messenger,
     private readonly OrgAccessSettings $settings,
-    private readonly AccountProxyInterface $currentUser,
   ) {}
 
   /**
@@ -140,9 +138,6 @@ class MassOrgAccessHooks {
       return;
     }
     $this->orgAccessChecker->populateOwnerGroupsFromCurrentUser($entity);
-    if (!$this->currentUser->isAnonymous()) {
-      $this->orgAccessChecker->applyDefaultsToNewEntity($entity, $this->currentUser);
-    }
   }
 
   /**
